requirements.txt
asgiref==3.2.3
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-storages==1.9.1
dropbox==9.4.0
emoji==0.5.4
future==0.18.2
gunicorn==20.0.0
hotxlfp==0.0.6
idna==2.8
lxml==4.4.1
pip-autoremove==0.9.1
ply==3.11
psycopg2==2.8.4
pycparser==2.19
pyngrok==2.0.3
python-dateutil==2.8.1
python-dotenv==0.10.3
python-telegram-bot==12.0.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.7
whitenoise==4.1.4
zipfile36==0.1.3
